Implement reading session
Now the users can read books and magazines.
In general I improve a lot of code measures and the creation of test users is with the dummyapi.io, for the books I want to search some api that has images and prices.
Remove the records becuz I think is a bad thing for my code, so i need to thing in a new way to refactor that repeat code.
Optimize abstract class and improve how are implemented on childs.
The next steps are implement the ads, reports and other things for a better ux.